Refactor: Deduplicate UI type definitions
Removed redundant type definitions in `ui/src/modTypes.ts` and updated UI components to use the shared types from `lib/readMods.ts`.
