+++ lib/readMods.ts
