+++ package.json
@@ -34,10 +34,10 @@
     "typescript": "^5.4.5"
   },
   "dependencies": {
-    "chalk": "^5.3.0",
+    "chalk": "^5.4.1",
     "clipboardy": "^4.0.0",
     "cors": "^2.8.5",
-    "dotenv": "^16.4.5",
+    "dotenv": "^16.5.0",
     "express": "^5.1.0",
     "jsonschema": "^1.5.0",
     "node-emoji": "^2.1.3",
